+++ src/serviceSatelliteClass.py
@@ -1,7 +1,7 @@
 import math
 
 class serviceSatellite:
-    def __init__(self, year, day, M):
+    def __init__(self, year, day, M, n = 1):
 
         # defines a circular geostationary orbit at a specified mean anomaly
         self.serviceSat = dict()
@@ -15,4 +15,4 @@ class serviceSatellite:
         self.serviceSat['eccentricity'] = 0 # circle
         self.serviceSat['omega'] = 0 # technically not defined, defaulted to 0
         self.serviceSat['M'] = math.radians(M) % (2 * math.pi) # Mean Anomaly, radians
-        self.serviceSat['n'] = 1 # Mean Motion, revs/day
+        self.serviceSat['n'] = n # Mean Motion, revs/day
